requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==5.0.0
billiard==4.2.1
boto3==1.35.55
botocore==1.35.55
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.2
django-rest-framework==0.1.0
django-storages==1.14.4
djangorestframework==3.15.2
jmespath==1.0.1
kombu==5.4.2
mysqlclient==2.2.5
pillow==11.0.0
prompt_toolkit==3.0.48
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.2.0
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
